+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.3] - 2026-01-04
+
+### Added
+- Added `Page()` method to router `Group` for fluent page registration API
+- Added `GroupPageBuilder` for building pages within route groups
+
+### Fixed
+- Fixed missing `Page()` method in router groups that broke example compilation
+
 ## [0.0.2] - 2026-01-04
 
 ### Changed
@@ -103,7 +112,8 @@ All releases include:
 
 ---
 
-[Unreleased]: https://github.com/xraph/forgeui/compare/v0.0.2...HEAD
+[Unreleased]: https://github.com/xraph/forgeui/compare/v0.0.3...HEAD
+[0.0.3]: https://github.com/xraph/forgeui/compare/v0.0.2...v0.0.3
 [0.0.2]: https://github.com/xraph/forgeui/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/xraph/forgeui/releases/tag/v0.0.1
 [0.1.0]: https://github.com/xraph/forgeui/releases/tag/v0.1.0